Add auto-zeroing feature
Passing the `-0` option to `get_region` will now add x- and y-offsets
that move the top-left of the total image to the origin (0,0). The
offset is calculated by scanning the names of every file in the `-i`
directory, which could be improved later.

Additionally, due to circular dependencies, code has been split into
smaller files, introducing `name.c/h` and `types.h`. `main.h` has been
removed.
