--- src/regions/main.h
@@ -1,8 +1,0 @@
-#ifndef _MAIN_H
-#define _MAIN_H
-
-struct options {
-    int filename_value_base;
-};
-
-#endif /* _MAIN_H */
