+++ src/regions/region.c
@@ -22,23 +22,18 @@
 #include <assert.h>
 
 #include "region.h"
-#include "main.h"
 #include "debug.h"
 #include "llist.h"
+#include "types.h"
+#include "name.h"
 
-struct tile {
-    struct region region;
-    char          filename[256];
-};
+int offset(struct region *reg, czi_coord_t x, czi_coord_t y) {
+    reg->up    += y;
+    reg->down  += y;
+    reg->left  += x;
+    reg->right += x;
 
-static bool tile_comparator(struct tile *a, struct tile *b) {
-    return !(a->region.up > b->region.up
-            || (a->region.up == b->region.up
-                && a->region.left > b->region.left));
-}
-
-static bool tile_ll_comparator(void *a, void *b) {
-    return tile_comparator((void *)a, (void *)b);
+    return 0;
 }
 
 void debug_region(const struct region *region) {
@@ -95,110 +90,12 @@ struct region *move_relative(struct region *root, struct region *x) {
     return x;
 }
 
-czi_coord_t str_czi_coord(char* str, char **end, int base) {
-    int errno_tmp = errno;
-    errno = 0;
-    czi_coord_t ret = strtol(str, end, base);
-    if (errno) {
-        err(errno, NULL);
-    }
-    errno = errno_tmp;
-    return ret;
-}
-
-/*
- * Like strstr(3), but points to the end of the needle rather than the start.
- */
-char *strstrend(const char *str, const char *needle) {
-    char *mid = strstr(str, needle);
-    if (!mid) {
-        return 0;
-    }
-    return mid + strlen(needle);
-}
-
-int set_side(struct dirent *ent, char *id, struct options *opts,
-        czi_coord_t *left, czi_coord_t *right, int *scale) {
-    char *filename = ent->d_name;
-    int base = opts->filename_value_base;
-
-    /*
-     * The following lines find the dimension, then the dimension's values. A
-     * problem could occur if a value hasn't been speicified for a dimension,
-     * in which case this would find the value of a different dimension.
-     *
-     * While the ordering of values in a filename has been agreed, this has
-     * been written to work with values in any order, just in case.
-     */
-    char *dim_start = strstrend(filename,id);
-    if (!dim_start) return 1;
-
-    char *left_start = strstrend(dim_start, "p");
-    if (!left_start) return 2;
-    *left = str_czi_coord(left_start, NULL, base);
-
-    if (right) {
-        char *size_start = strstrend(dim_start, "s");
-        if (!size_start) return 3;
-        *right = *left + str_czi_coord(size_start, NULL, base);
-    }
-
-    if (scale) {
-        char *scale_start = strstrend(dim_start, "r");
-        if (!scale_start) return 4;
-
-        czi_coord_t scale_tmp = str_czi_coord(scale_start, NULL, base);
-        assert(scale_tmp <= INT_MAX);
-        *scale = (int) scale_tmp;
-    }
-
-    return 0;
-}
-
-int get_region(struct dirent *ent, struct region *buf, struct options *opts) {
-    int ss_ret;
-
-    ss_ret = set_side(ent, "X", opts, &(buf->left), &(buf->right),
-            &(buf->scale));
-    if (ss_ret) return ss_ret;
-
-    int scale_tmp;
-    ss_ret = set_side(ent, "Y", opts, &(buf->up), &(buf->down), &scale_tmp);
-    if (ss_ret) return ss_ret;
-    assert(buf->scale == scale_tmp);
-
-    return 0;
-}
-
 void print_tiles(llist *list) {
     debug("%s\n", "--");
     for (struct ll_node *node = list; node; node = node->next) {
         debug("%s\n", ((struct tile *) (node->content))->filename);
     }
     debug("%s\n", "--");
-}
-
-llist *find_relevant_tiles(struct region *desired, char *tile_dirname,
-        struct options *opts) {
-    llist *included_tiles = NULL;
-    DIR *dir = opendir(tile_dirname);
-    if (!dir) {
-        err(errno, "%s", tile_dirname);
-    }
-
-    struct dirent *ent;
-    while ((ent = readdir(dir))) {
-        struct region tile_region;
-        if (get_region(ent, &tile_region, opts)) continue;
-        if (tile_region.scale != desired->scale) continue;
-        if (overlaps(&tile_region, desired)) {
-            struct tile *tile = (struct tile *) malloc(sizeof (*tile));
-            tile->region = tile_region;
-            strncpy(tile->filename, ent->d_name, strlen(ent->d_name)+1);
-            included_tiles = ll_add_item(included_tiles, tile, &tile_ll_comparator);
-        }
-    }
-    return included_tiles;
 }
 
 struct filenamedata {
